+++ CHANGELOG.md
@@ -358,3 +358,4 @@
 * Removed UI folder (remaining contents moved to Shared)
 * Added speaker name and better color handling to SubtitleUtilsEx
 * Fixed PostProcessVolume layer colliding with other layers
+* Allow and attempt impossible skill checks can now be overridden with dialogue scene/frame options